feat: complete template

+++ scripts/utils.js
@@ -7,5 +7,9 @@ exports.clearFile = function clearFile(filePath) {
 
 exports.clearFolder = function clearFolder(folderPath) {
   if (!fs.existsSync(folderPath)) return
+  if (!fs.rmdirSync) {
+    console.warn('installed nodejs is outdated, visit https://nodejs.org/ to update')
+    process.exit(1)
+  }
   fs.rmdirSync(folderPath, { recursive: true })
 }
